+++ coverage/parser.py
@@ -654,7 +654,7 @@ class NodeList(ast.AST):
 # TODO: Shouldn't the cause messages join with "and" instead of "or"?
 
 
-def is_constant_test_expr(node: ast.AST) -> bool:
+def is_constant_test_expr(node: ast.AST) -> tuple[bool, bool]:
     """Is this a compile-time constant test expression?
 
     We don't try to mimic all of CPython's optimizations.  We just have to
@@ -662,15 +662,20 @@ def is_constant_test_expr(node: ast.AST) -> bool:
 
     """
     if isinstance(node, ast.Constant):
-        return True
+        return True, bool(node.value)
     elif isinstance(node, ast.Name):
         if node.id in ["True", "False", "None", "__debug__"]:
-            return True
+            return True, eval(node.id)  # pylint: disable=eval-used
     elif isinstance(node, ast.UnaryOp) and isinstance(node.op, ast.Not):
-        return is_constant_test_expr(node.operand)
+        is_constant, val = is_constant_test_expr(node.operand)
+        return is_constant, not val
     elif isinstance(node, ast.BoolOp):
-        return all(is_constant_test_expr(v) for v in node.values)
-    return False
+        rets = [is_constant_test_expr(v) for v in node.values]
+        is_constant = all(is_const for is_const, _ in rets)
+        if is_constant:
+            op = any if isinstance(node.op, ast.Or) else all
+            return True, op(v for _, v in rets)
+    return False, False
 
 
 class AstArcAnalyzer:
@@ -1156,10 +1161,14 @@ class AstArcAnalyzer:
 
     def _handle__If(self, node: ast.If) -> set[ArcStart]:
         start = self.line_for_node(node.test)
-        from_start = ArcStart(start, cause="the condition on line {lineno} was never true")
-        exits = self.process_body(node.body, from_start=from_start)
-        from_start = ArcStart(start, cause="the condition on line {lineno} was always true")
-        exits |= self.process_body(node.orelse, from_start=from_start)
+        constant_test, val = is_constant_test_expr(node.test)
+        exits = set()
+        if not constant_test or val:
+            from_start = ArcStart(start, cause="the condition on line {lineno} was never true")
+            exits |= self.process_body(node.body, from_start=from_start)
+        if not constant_test or not val:
+            from_start = ArcStart(start, cause="the condition on line {lineno} was always true")
+            exits |= self.process_body(node.orelse, from_start=from_start)
         return exits
 
     if sys.version_info >= (3, 10):
@@ -1271,7 +1280,7 @@ class AstArcAnalyzer:
 
     def _handle__While(self, node: ast.While) -> set[ArcStart]:
         start = to_top = self.line_for_node(node.test)
-        constant_test = is_constant_test_expr(node.test)
+        constant_test, _ = is_constant_test_expr(node.test)
         top_is_body0 = False
         if constant_test:
             top_is_body0 = True